adaptive-diagnostic-engine/requirements.txt
fastapi
uvicorn
pymongo
openai
python-dotenv
pydantic
prometheus-fastapi-instrumentator
